% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_appid_config.R
\name{gen_appid_config}
\alias{gen_appid_config}
\title{Generate a configuration file}
\usage{
gen_appid_config(name = NULL)
}
\arguments{
\item{name}{App's name}
}
\value{
Configuration file
}
\description{
Generate a configuration file
}
