% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_user_info}
\alias{get_user_info}
\title{Get user information}
\usage{
get_user_info(input, output, session)
}
\arguments{
\item{input}{Input}

\item{output}{Output}

\item{session}{Session}
}
\value{
Name
}
\description{
This is a shiny module to get user information
}
