% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oauth_config}
\alias{oauth_config}
\title{Get information from yaml file}
\usage{
oauth_config(file)
}
\arguments{
\item{file}{A character string naming a file}
}
\value{
A list with input information from yaml file
}
\description{
Get information from yaml file
}
