% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyAppId}
\alias{shinyAppId}
\title{Modified shinyApp}
\usage{
shinyAppId(ui, server)
}
\arguments{
\item{ui}{The UI definition of the app}

\item{server}{A function with three parameters: input, output, and session. The function is called once for each session ensuring that each app is independent}
}
\value{
An object that represents the app
}
\description{
Modified shinyApp
}
